requirements.txt
# pip3 install -r requirements.txt
airtest==1.2.6
airtest-selenium==1.0.3
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
cached-property==1.5.2
certifi==2022.5.18.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
comtypes==1.1.11
cryptography==37.0.2
ddt==1.5.0
decorator==5.1.1
Deprecated==1.2.13
et-xmlfile==1.1.0
facebook-wda==1.4.6
h11==0.13.0
hrpc==1.0.9
idna==2.10
iniconfig==1.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
MouseInfo==0.1.3
mss==6.1.0
numpy==1.22.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.10
outcome==1.1.0
packaging==21.3
Pillow==9.1.1
pip==22.0.4
pluggy==1.0.0
pocoui==1.0.87
py==1.11.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.1.2
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-rerunfailures==10.2
pytweening==1.0.4
pywin32==304
pywinauto==0.6.3
PyYAML==6.0
requests==2.27.1
retry==0.9.2
selenium==4.2.0
setuptools==58.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
TestRunner==0.8.4
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
uiautomator==1.0.2
unittestreport==1.5.0
urllib3==1.26.9
websocket-client==0.48.0
wrapt==1.14.1
wsproto==1.1.0